netlink: collapse integration tests into one file, run with sudo in travis
