+++ .travis.yml
@@ -14,4 +14,6 @@ script:
   - go vet ./...
   - staticcheck ./...
   - golint -set_exit_status ./...
-  - go test -v -race -tags=integration ./...
+  - go test -v -race ./...
+  - go test -c -race -tags=integration .
+  - sudo ./netlink.test -test.v -test.run TestLinuxConnIntegration*
